Update README.md
In this project, we explore and analyze a dataset containing emergency 911 calls in
Montgomery County, PA. The dataset provides information about the reasons for the calls,
the location, and the timestamp of each call. The goal of this project is to gain insights into
the patterns, trends, and distribution of emergency calls and understand the primary reasons
for these calls.